[soham]Added:Applied changes to implement SQL

+++ EmployeePayrollService/EmployeeModel.cs
@@ -14,9 +14,9 @@ namespace EmployeePayrollService
         public string Gender { get; set; }
         public decimal BasicPay { get; set; }
         public double Deductions { get; set; }
-        public double TaxablePay { get; set; }
-        public double Tax { get; set; }
-        public double NetPay { get; set; }
+        public decimal TaxablePay { get; set; }
+        public decimal Tax { get; set; }
+        public decimal NetPay { get; set; }
         public DateTime StartDate { get; set; }
     }
 }
